refactor(precompile): create crypto submodule for blake2
Move the BLAKE2b compression algorithm (`algo` module with SIGMA,
IV, G function, compress, round) and the AVX2 optimized variant
into a dedicated `crypto.rs` submodule, separating pure
cryptographic logic from the precompile runner.

Closes part of #2782
